Resizing bug when scrolling
Avoid resizing container when scrolling far past the Masonry on a page with many elements below the Masonry. `this.masonryContainer.getBoundingClientRect().top` becomes a very low negative number when scrolling past the Masonry. This, negative, number seems to be irrelevant to the calculation of the height of the Masonry. So perhaps it's enough to just check if the value is negative, and skip the setting of the height at this point.